fix(setup): avoid registering services per entity
Currently, when multiple entities are registered, the two engine services are registered for each entity. This causes whichever vehicle is registered last to be the only one that the services work for. This change registers the services once in `async_setup` and utilizes the `entity_id` to determine which vehicle the service should be called on.